changed ints to dbl, explorer complete

+++ Account.cs
@@ -3,6 +3,6 @@ namespace FirstBank
   public class Account
   {
     public string AccountName { get; set; }
-    public int AccountAmount { get; set; }
+    public double AccountAmount { get; set; }
   }
 }
